--- apps/dapp/src/components/Pages/Core/DappPages/Dashboard/Gnosis.tsx
@@ -1,5 +1,0 @@
-const Gnosis = () => {
-  return <h1>Gnosis</h1>;
-};
-
-export default Gnosis;
