--- apps/dapp/src/components/Pages/Core/DappPages/Dashboard/Ramos.tsx
@@ -1,5 +1,0 @@
-const Ramos = () => {
-  return <h1>Ramos</h1>;
-};
-
-export default Ramos;
