--- apps/dapp/src/components/Pages/Core/DappPages/Dashboard/Tlc.tsx
@@ -1,5 +1,0 @@
-const Tlc = () => {
-  return <h1>Tlc</h1>;
-};
-
-export default Tlc;
